prpr/requirements.txt
PyYAML==5.4.1
loguru==0.5.3
yandex_tracker_client==1.9
